Add a timezone API.
Add a small timezone API that returns the UTC offset, timezone name, and
daylight saving time status for a given `datetime` timestamp.

This is not a full timezone database API. It's designed so that
implementations that don't wish to do anything with timezones can just
support a single "UTC" time zone, which can be easily implemented.

It's also not a time formatting or locale API. It doesn't say how to
format a human-readable timestamp; it just provides the information that
would typically be formatted into a human-readable timestamp.

This API expects a `timezone` handle argument will be added to the
command entrypoint function, providing programs with a `timezone` to
provide timezone information about wall-clock timestamps.

Fixes #25.