Add a polling interface for monotonic clocks. (#34)
The new wasi-poll API has a `pollable` type. Add subscribe methods to
the monotonic clock types to allow polling for timeouts.

This adds a dependency on wasi-io, and is currently implemented in the
repository in a low-tech way, which allows the generated documentation
to work, though it's likely to evolve.