remove timezone interface from the proposal
We believe there is value in providing timezone support in this package,
but for the sake of shipping Preview 2, we are out of time to implement
these methods and test a a cross platform implementation of these in two
separate engines.

We intend for this interface to return to the wasi:clocks package and
be made in a post-Preview 2 patch version 0.2.n. Re-introducing this
interface may be done with a revert of this commit once we have shipped
Preview 2 and figured out the process for shipping patch versions.